Integrata correzione "Remove exception handling from extractArchive"
https://github.com/junrar/junrar/pull/2

As the catches didn't permit to handle correctly any of the exceptions
thrown during the process of extraction, it is better to throw them to
the caller. At this level, it is not possible to know the handling
strategy to use and this code should not handle them at all.

alecharp