Integrata aggiunta "Extracting InputStream directly" di daniel-rabe
https://github.com/edmund-wagner/junrar/pull/34

I did some changes to JunRar to extract InputStreams directly, without
the need to write it in a File first.

My project loads files from websides (e.g. RAR-Files), the results are
InputStreams of that files. When I want to know what is inside a
RAR-File I don't want to write that InputStream in a File everytime, so
I made this little changes to junrar.

greets Daniel Rabe