In fact, it seems to be a bug of the implementation. This is the workaround I have used in order to avoid the problem:
In com.github.junrar.unpack.ppm.ModelPPM.java, line 196: MaxMB = unpackRead.getChar();

the method getChar, in some weird situations, return a very big number. It should be due to a broken header or a header option not supported by junrar.

My workaround was to check if MaxMB was greater than 1 and set to 1. I've been using this fix for a long time without problems.